requirements.txt
# Core dependencies
reportlab==4.0.7
PyPDF2==3.0.1

# AI providers
openai==1.12.0
anthropic==0.18.1

# Utilities
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.3
